Including implementation of Simulated Annealing, GRASP, and more solution obtained with Gurobi.

+++ src/types/item.h
@@ -1,5 +1,5 @@
 typedef struct {
     int id;          // Item id
-    double weight;   // Item weight
-    double value;    // Item value
+    int weight;   // Item weight
+    int value;    // Item value
 } Item;
